Moved table selection functions into operations.
