+++ db/columns.py
@@ -15,7 +15,7 @@ from psycopg2.errors import InvalidTextRepresentation, InvalidParameterValue
 from db import constants
 from db.constraints import operations as constraint_operations
 from db.constraints import utils as constraint_utils
-from db.tables import utils as table_utils
+from db.tables.operations.select import get_oid_from_table, reflect_table_from_oid
 from db.types import alteration
 from db.types.base import get_db_type_name
 from db.utils import execute_statement
@@ -158,7 +158,7 @@ class MathesarColumn(Column):
                 and self.table_ is not None
                 and inspect(self.engine).has_table(self.table_.name, schema=self.table_.schema)
         ):
-            table_oid = table_utils.get_oid_from_table(
+            table_oid = get_oid_from_table(
                 self.table_.name, self.table_.schema, self.engine
             )
             return get_column_index_from_name(
@@ -170,7 +170,7 @@ class MathesarColumn(Column):
     @property
     def default_value(self):
         if self.table_ is not None:
-            table_oid = table_utils.get_oid_from_table(
+            table_oid = get_oid_from_table(
                 self.table_.name, self.table_.schema, self.engine
             )
             return get_column_default(table_oid, self.column_index, self.engine)
@@ -245,7 +245,7 @@ def create_column(engine, table_oid, column_data):
         logger.warning("Requested type not supported. falling back to VARCHAR")
         sa_type = supported_types["VARCHAR"]
         column_type_options = {}
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
 
     try:
         column = MathesarColumn(
@@ -258,7 +258,7 @@ def create_column(engine, table_oid, column_data):
         else:
             raise e
 
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     try:
         with engine.begin() as conn:
             ctx = MigrationContext.configure(conn)
@@ -273,7 +273,7 @@ def create_column(engine, table_oid, column_data):
             raise e
 
     return get_mathesar_column_with_engine(
-        table_utils.reflect_table_from_oid(table_oid, engine).columns[column_data[NAME]],
+        reflect_table_from_oid(table_oid, engine).columns[column_data[NAME]],
         engine
     )
 
@@ -285,7 +285,7 @@ def alter_column(engine, table_oid, column_index, column_data):
     DEFAULT_KEY = 'default_value'
     NAME_KEY = NAME
 
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     column_index = int(column_index)
 
     with engine.begin() as conn:
@@ -306,7 +306,7 @@ def alter_column(engine, table_oid, column_index, column_data):
             rename_column(table, column_index, engine, conn, name)
 
     return get_mathesar_column_with_engine(
-        table_utils.reflect_table_from_oid(table_oid, engine).columns[column_index],
+        reflect_table_from_oid(table_oid, engine).columns[column_index],
         engine
     )
 
@@ -418,7 +418,7 @@ def _batch_alter_table_columns(table, column_data_list, connection):
 
 
 def batch_update_columns(table_oid, engine, column_data_list):
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     _validate_columns_for_batch_update(table, column_data_list)
     with engine.begin() as conn:
         _batch_update_column_types(table, column_data_list, conn, engine)
@@ -427,7 +427,7 @@ def batch_update_columns(table_oid, engine, column_data_list):
 
 def drop_column(table_oid, column_index, engine):
     column_index = int(column_index)
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     column = table.columns[column_index]
     with engine.begin() as conn:
         ctx = MigrationContext.configure(conn)
@@ -436,7 +436,7 @@ def drop_column(table_oid, column_index, engine):
 
 
 def get_column_default(table_oid, column_index, engine, connection_to_use=None):
-    table = table_utils.reflect_table_from_oid(table_oid, engine, connection_to_use)
+    table = reflect_table_from_oid(table_oid, engine, connection_to_use)
     column = table.columns[column_index]
     if column.server_default is None:
         return None
@@ -509,7 +509,7 @@ def compile_copy_column(element, compiler, **_):
 
 
 def duplicate_column_data(table_oid, from_column, to_column, engine):
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     copy = CopyColumn(
         table.schema,
         table.name,
@@ -527,7 +527,7 @@ def duplicate_column_data(table_oid, from_column, to_column, engine):
 
 def duplicate_column_constraints(table_oid, from_column, to_column, engine,
                                  copy_nullable=True):
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     if copy_nullable:
         with engine.begin() as conn:
             change_column_nullable(table, to_column, engine, conn, table.c[from_column].nullable)
@@ -547,7 +547,7 @@ def duplicate_column(
     table_oid, copy_from_index, engine,
     new_column_name=None, copy_data=True, copy_constraints=True
 ):
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     from_column = table.c[copy_from_index]
     if new_column_name is None:
         new_column_name = _gen_col_name(table, from_column.name)
@@ -577,6 +577,6 @@ def duplicate_column(
             copy_nullable=copy_data
         )
 
-    table = table_utils.reflect_table_from_oid(table_oid, engine)
+    table = reflect_table_from_oid(table_oid, engine)
     column_index = get_column_index_from_name(table_oid, new_column_name, engine)
     return get_mathesar_column_with_engine(table.c[column_index], engine)
