+++ db/tables/operations/__init__.py
@@ -2,5 +2,6 @@ from db.tables.operations.create import *  # noqa
 from db.tables.operations.drop import *  # noqa
 from db.tables.operations.merge import *  # noqa
 from db.tables.operations.move_columns import *  # noqa
+from db.tables.operations.select import *  # noqa
 from db.tables.operations.split import *  # noqa
 from db.tables.operations.update import *  # noqa
