+++ db/tests/constraints/operations/test_drop.py
@@ -1,7 +1,7 @@
 from sqlalchemy import String, Integer, Column, Table, MetaData
 
 from db.constraints.operations import create_unique_constraint, drop_constraint
-from db.tables import utils as table_utils
+from db.tables.operations.select import get_oid_from_table, reflect_table_from_oid
 from db.tests.constraints import utils as test_utils
 
 
@@ -17,12 +17,12 @@ def test_drop_unique_constraint(engine_with_schema):
     )
     table.create()
 
-    table_oid = table_utils.get_oid_from_table(table_name, schema, engine)
+    table_oid = get_oid_from_table(table_name, schema, engine)
     create_unique_constraint(table.name, schema, engine, [unique_column_name])
-    altered_table = table_utils.reflect_table_from_oid(table_oid, engine)
+    altered_table = reflect_table_from_oid(table_oid, engine)
     test_utils.assert_primary_key_and_unique_present(altered_table)
     unique_constraint = test_utils.get_first_unique_constraint(altered_table)
 
     drop_constraint(table_name, schema, engine, unique_constraint.name)
-    new_altered_table = table_utils.reflect_table_from_oid(table_oid, engine)
+    new_altered_table = reflect_table_from_oid(table_oid, engine)
     test_utils.assert_only_primary_key_present(new_altered_table)
