+++ db/tests/tables/operations/test_update.py
@@ -2,8 +2,9 @@ import pytest
 from sqlalchemy.exc import NoSuchTableError
 
 from db.tables.operations.create import create_mathesar_table
+from db.tables.operations.select import get_oid_from_table
 from db.tables.operations.update import rename_table
-from db.tables import utils as table_utils
+from db.tables.utils import reflect_table
 from db.tests.tables import utils as test_utils
 
 
@@ -12,17 +13,17 @@ def test_rename_table(engine_with_schema):
     table_name = "test_rename_table"
     new_table_name = "test_rename_table_new"
     old_table = create_mathesar_table(table_name, schema, [], engine)
-    old_oid = table_utils.get_oid_from_table(old_table.name, old_table.schema, engine)
+    old_oid = get_oid_from_table(old_table.name, old_table.schema, engine)
 
     rename_table(table_name, schema, engine, new_table_name)
-    new_table = table_utils.reflect_table(new_table_name, schema, engine)
-    new_oid = table_utils.get_oid_from_table(new_table.name, new_table.schema, engine)
+    new_table = reflect_table(new_table_name, schema, engine)
+    new_oid = get_oid_from_table(new_table.name, new_table.schema, engine)
 
     assert old_oid == new_oid
     assert new_table.name == new_table_name
 
     with pytest.raises(NoSuchTableError):
-        table_utils.reflect_table(table_name, schema, engine)
+        reflect_table(table_name, schema, engine)
 
 
 def test_rename_table_foreign_key(engine_with_schema):
@@ -36,6 +37,6 @@ def test_rename_table_foreign_key(engine_with_schema):
 
     rename_table(table_name, schema, engine, new_table_name)
 
-    related_table = table_utils.reflect_table(related_table_name, schema, engine)
+    related_table = reflect_table(related_table_name, schema, engine)
     fk = list(related_table.foreign_keys)[0]
     assert fk.column.table.name == new_table_name
