+++ db/tests/types/test_alteration.py
@@ -9,7 +9,7 @@ from sqlalchemy.exc import DataError
 
 from db import types, columns
 from db.tables.operations.create import create_mathesar_table
-from db.tables import utils as table_utils
+from db.tables.operations.select import get_oid_from_table
 from db.tests.types import fixtures
 from db.types import alteration
 from db.types.base import PostgresType, MathesarCustomType, get_qualified_name, get_available_types
@@ -554,7 +554,7 @@ def test_alter_column_casts_data_gen(
         res = conn.execute(sel).fetchall()
     actual_value = res[0][0]
     assert actual_value == out_val
-    table_oid = table_utils.get_oid_from_table(TABLE_NAME, schema, engine)
+    table_oid = get_oid_from_table(TABLE_NAME, schema, engine)
     actual_default = columns.get_column_default(table_oid, 0, engine)
     assert actual_default == out_val
