+++ db/types/alteration.py
@@ -3,7 +3,7 @@ from sqlalchemy.sql import quoted_name
 from sqlalchemy.sql.functions import Function
 
 from db import columns
-from db.tables import utils as table_utils
+from db.tables.operations.select import get_oid_from_table, reflect_table_from_oid
 from db.types import base, email
 from db.utils import execute_statement
 
@@ -110,9 +110,9 @@ def alter_column_type(
     target_type = supported_types.get(target_type_str)
     schema = table.schema
 
-    table_oid = table_utils.get_oid_from_table(table.name, schema, engine)
+    table_oid = get_oid_from_table(table.name, schema, engine)
     # Re-reflect table so that column is accurate
-    table = table_utils.reflect_table_from_oid(table_oid, engine, connection)
+    table = reflect_table_from_oid(table_oid, engine, connection)
     column = table.columns[column_name]
     column_index = columns.get_column_index_from_name(table_oid, column_name, engine, connection)
